big commit: change model inheritance & constructor pattern
- model:
    my.M        --->    model value
    my.model    --->    model constructor

+ constructor pattern:

Multiple patterns are possible, either by calling `fst` directly
or by supplying the array of arguments in the branches array of the
parent node.
```
    fst('tag', M => [...branches(M)])

    [ 'tag', M => [...branches(M)] ]
```

+ my.model(f, ...fs) :

Customize the way a node should inherit the model from its parent.
Each of the fs is of type `Object || Model -> Object`.
If `f :: Model -> Object` is a function,
keys will be updated from f(M).
If `f :: Object` this is equivalent to calling:
 `my.model(__.id, f,...fs)`