upgrade IO methods to accept map nodes
ie  `IO.set(circle.map())([...ms])` will update all instances,
    `IO.place(circle.map())([...ms])` will place or replace nodes
    where appropriate.

However there is now way to remove/place a node at the end,
as the indexes are infered from the array's.
This would work with records, yet we lose the ordering.

The effect would be that of calling, say:
    `IO.replace(circle.place(['circle', i]))(mi);

We could do:
    `IO.set(circles, [i, j, k])(ms)` to explicitly map those to indices,
    or override the pull function to accept an array of indices
    as secondary argument:
    `IO.set(circles)(ms, [i, j, k])`