restore IO to a cleanly working state
IO operations accept only single nodes as arguments.
Trying to be type agnostic to please one's laziness
is python-like and makes the code messy!

Instead focus of linking IO methods to dom instances,
as it is a pushforward:

    node.io('put')
    node.io('replace')

    nodes.io('remove', [0, 2, 12])

As this will be much cleaner! Just a figure a clever way to chain these:

    m => nodes.io('set', [0, 1, 3])(m)
        .bind(nodes.io('remove', [2, 4, 5])
<=>
    nodes
        .io('set',      [0, 1, 3])
        .io('remove',   [2, 4, 5])
        .io('replace',  [6, 7, 8]);

This is especially usefull for map nodes. Instead of array indices,
one could provide with a filtering function.