point branch to 'Dom(a)' instead 'a -> Node'
This is a big change though almost seamless thanks
to the refactoring of `data.js`.

Instead of viewing dom instances as `a -> Node` functions,
it is really much cleaner to treat them internally as what
they are, i.e. this `a`-dependent tree-like object `self`:

    Dom(a) = Tree(a, Dom.node(a))

Then `data.js` simply implements the natural map `Tree.apply`
with a bit of sugar:

    data : Dom(a) -> Data = Tree(d)

And finally, `DOM.js` maps `Tree(d)` to `Node`.